requirements.txt
# Main packages
Flask~=2.1.2
flask-restx
Flask-CORS
Flask-SQLAlchemy
marshmallow~=3.16.0
PyJWT~=2.4.0

# Docker
alembic==1.7.5
click==8.0.3
Flask-Migrate==3.1.0
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2-binary==2.9.2

# For test
pytest~=7.1.2
pytest-cov
SQLAlchemy~=1.4.36
Werkzeug~=2.1.2